Add a "room" label to each thermostat timeseries
This adds a new "room" label to each timeseries associated
with a thermostat. The value matches what it displayed in
the Google Home app. Google Home app assigns each thermostat
to awq room.

The "room" label is different from the pre-existing "label"
label which is set to the Label field of the Where group of
settings in the Nest app.