Omit nest_setpoint_temperature_celsius when no value
Prior to this commit, when the thermostat does not have a
set point for heating, such as when it's in the OFF or COOL
mode, the nest_setpoint_temperature_celsius metric was output
as 0. This is problematic because it may be interpreted as the
target temperature being set to 0.

This commit fixes this ambiguity by not outputting the
nest_setpoint_temperature_celsius when no set point is set.
This happens, for example, when you switch the thermostat into
the Off mode.